requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.65.0
anyio==4.10.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
certifi==2025.8.3
chainlit==2.7.2
charset-normalizer==3.4.3
chevron==0.14.0
click==8.2.1
colorama==0.4.6
cuid==0.4
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
fastapi==0.116.1
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
googleapis-common-protos==1.70.0
groq==0.31.0
grpcio==1.74.0
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
Lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.13.1
monotonic==1.6
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
openai==1.105.0
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-alephalpha==0.46.2
opentelemetry-instrumentation-anthropic==0.46.2
opentelemetry-instrumentation-bedrock==0.46.2
opentelemetry-instrumentation-chromadb==0.46.2
opentelemetry-instrumentation-cohere==0.46.2
opentelemetry-instrumentation-crewai==0.46.2
opentelemetry-instrumentation-google-generativeai==0.46.2
opentelemetry-instrumentation-groq==0.46.2
opentelemetry-instrumentation-haystack==0.46.2
opentelemetry-instrumentation-lancedb==0.46.2
opentelemetry-instrumentation-langchain==0.46.2
opentelemetry-instrumentation-llamaindex==0.46.2
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-marqo==0.46.2
opentelemetry-instrumentation-mcp==0.46.2
opentelemetry-instrumentation-milvus==0.46.2
opentelemetry-instrumentation-mistralai==0.46.2
opentelemetry-instrumentation-ollama==0.46.2
opentelemetry-instrumentation-openai==0.46.2
opentelemetry-instrumentation-openai-agents==0.46.2
opentelemetry-instrumentation-pinecone==0.46.2
opentelemetry-instrumentation-qdrant==0.46.2
opentelemetry-instrumentation-redis==0.57b0
opentelemetry-instrumentation-replicate==0.46.2
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-sagemaker==0.46.2
opentelemetry-instrumentation-sqlalchemy==0.57b0
opentelemetry-instrumentation-threading==0.57b0
opentelemetry-instrumentation-together==0.46.2
opentelemetry-instrumentation-transformers==0.46.2
opentelemetry-instrumentation-urllib3==0.57b0
opentelemetry-instrumentation-vertexai==0.46.2
opentelemetry-instrumentation-watsonx==0.46.2
opentelemetry-instrumentation-weaviate==0.46.2
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.57b0
packaging==25.0
posthog==3.25.0
propcache==0.3.2
protobuf==6.32.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.3
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
tomli==2.2.1
torch==2.8.0
tqdm==4.67.1
traceloop-sdk==0.46.2
transformers==4.56.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==0.24.0
wrapt==1.17.3
wsproto==1.2.0
yarl==1.20.1
zipp==3.23.0
streamlit==1.49.1
plotly==6.3.0
pymongo
